requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.3
python-dotenv==1.0.0
google-generativeai>=0.8.5
google-generativeai>=0.8.5